chore: split cli and functions
Split the functions out to __init__.py and introduce a class for the
application, improving separation of concerns.

This allows for users to simply import our module and call
the functions as they desire, with the CLI being optional.

Additionally, if someone doesn't like the way that we handle
configuration, they can import our module instead of forking the
project.

This also makes the code slightly more readable, as well as reducing the
number of variables that are globally-scoped in the file.